pyproject.toml
[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"

[project]
name = "ml_ops_project"
version = "0.0.1"
description = "A short description of the project."
authors = [
  { name = "Team38", email = "your@email.com" },
]

keywords = ["machine learning", "MLOps"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
]
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "numpy<2",
    "datasets>=4.4.2",
    "fastapi==0.115.6",
    "ipykernel>=7.1.0",
    "lightning[pytorch-extra]>=2.6.0",
    "scikit-learn>=1.8.0",
    "tensorboard>=2.20.0",
    "torch==2.2.2",
    "torchvision",
    "transformers>=4.57.3",
    "typer==0.15.1",
    "uvicorn==0.34.0",
    "wandb>=0.23.1",
    "pytest-cov>=6.3.0",
    "hydra-core>=1.3.2",
    "dvc>=3.66.1",
    "dvc-gs>=3.0.2",
    "onnx>=1.20.1",
    "onnxruntime>=1.23.2",
    "onnxscript>=0.5.7",
    "netron>=8.8.3",
]

[dependency-groups]
dev = [
    "coverage==7.6.9",
    "invoke==2.2.0",
    "mkdocs-material==9.4.6",
    "mkdocs==1.6.1",
    "pre-commit==4.1.0",
    "pytest==8.3.4",
    "ruff==0.1.3",
    "mkdocstrings-python==1.12.2",
    "gcloud>=0.18.3",
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.uv.sources]
torch = [
    { index = "pytorch-cpu" },
]
torchvision = [
    { index = "pytorch-cpu" },
]

[tool.uv]
required-environments = [
    "sys_platform == 'linux' and platform_machine == 'x86_64'",
]
[tool.ruff]
line-length = 120
target-version = "py311"
extend-exclude = ["notebooks"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "TID", # flake8-tidy-imports
    "ARG", # flake8-unused-arguments
    "RUF", # Ruff-specific rules
]

[tool.coverage.run]
omit = ["tests/*"]
